src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/attribute/reference/CollectionStringAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.attribute.CollectionAttributeParser;
import org.openqa.selenium.WebElement;

public class CollectionStringAttributeParser extends CollectionAttributeParser {

    protected CollectionStringAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return element.getAttribute(super.attribute);
    }

}
