src/test/java/com/github/brunomndantas/jscrapper/support/parser/collection/attribute/reference/CollectionStringAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.collection.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class CollectionStringAttributeParserTest{


    @Test
    public void constructorTest() {
        String attribute = "";
         CollectionStringAttributeParser parser = new  CollectionStringAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
         CollectionStringAttributeParser parser = new  CollectionStringAttributeParser("");

        assertEquals("text", parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "text";
            }
        }));
    }

}
